+++ src/widget/select_tool.rs
@@ -29,7 +29,6 @@ impl SelectToolWidget {
         let mut buttons = vec![
             ButtonWidget::new(ButtonKind::Basic, IconId::Select),
             ButtonWidget::new(ButtonKind::Basic, IconId::Lasso),
-            ButtonWidget::new(ButtonKind::Basic, IconId::SelectBucket),
         ];
         buttons[tool_to_index(current)].set_kind(ButtonKind::BasicPressed);
         Ok(Self {
@@ -111,7 +110,6 @@ fn tool_to_index(tool: SelectTool) -> usize {
     match tool {
         SelectTool::Rectangle => 0,
         SelectTool::Lasso => 1,
-        SelectTool::Bucket => 2,
     }
 }
 
@@ -119,7 +117,6 @@ fn icon_to_tool(icon: IconId) -> Result<SelectTool> {
     match icon {
         IconId::Select => Ok(SelectTool::Rectangle),
         IconId::Lasso => Ok(SelectTool::Lasso),
-        IconId::SelectBucket => Ok(SelectTool::Bucket),
         _ => Err(orfail::Failure::new("unreachable")),
     }
 }